requirements.txt
streamlit
streamlit-audio-recorder
sounddevice
wavio
SpeechRecognition
gTTS
python-dotenv
google-generativeai
pydub